fix(ci): Fail the action pin check when it checked nothing
Addresses the review of #116. The scripts were ported from a repository with
a different lint, Prettier and module layout, and carried several artefacts of
that move plus one substantive hole.

The hole: with an invalid or rate-limited token every reference resolved to
nothing, all 29 were reported unknown, and the process still exited 0. A check
that checked nothing was indistinguishable from a clean run. classifyPin now
separates the two cases - "unresolved" for a pin that carries both a SHA and a
"# <tag>" comment and so was owed a comparison that did not happen, and
"unknown" for a pin with nothing to compare against - and a new summarize()
decides the exit code from the tally. Stale or unresolved fails; unknown does
not. The summary line now states resolved vs. unresolved counts outright.

A quoted uses: "owner/repo@sha" value previously mis-parsed the owner as
'"actions', which degraded to exactly that silent unknown + exit 0. The parser
now peels a quoted scalar before looking for the trailing comment.

The two eslint-disable comments for security/detect-non-literal-fs-filename
broke npm run lint outright ("Definition for rule ... was not found"):
eslint-plugin-security is registered for src/**/*.js only in this repository,
so the rule name is undefined under scripts/. Dropped rather than widening the
plugin's scope - eslint.config.mjs deliberately holds published library code to
a stricter bar than tooling, and extending that is a policy change this PR has
no business making. The justification for the fs reads is kept as prose on
collectPins.

Both files also failed npm run format:check; reformatted with the repository's
own Prettier (trailing commas, arrow params without parens).

Renamed .mjs to .js. package.json sets "type": "module", so a .js file here is
already ESM, and the repository's tooling is keyed to the .js extension:
.lintstagedrc.json only globs *.js, and Jest's default transform only matches
\.[jt]sx?$, so an .mjs module cannot be unit-tested without extra config.

Header comments corrected - they claimed a tests/ directory (it is test/), an
npm run security:action-pins script that does not exist, a CommonJS rewrite
(the file is ESM) referencing a scripts/check-sarif.js that does not exist, and
a main-branch pin for Dependency-Check_Action that this repository does not
have (it is pinned at 1.1.0). The scheduled-automation ban is #103 here; #45 is
an unrelated eslint bump.

The workflow_dispatch job inherited security-events: write from the workflow
default. It uploads no SARIF, so it is narrowed to contents: read.
